chore: fix circular dependencies.
